GP131: Update TEE_InitRefAttribute()
Update TEE_InitRefAttribute() according to TEE Internal Core API version
1.3.1. The size parameter is changed to use size_t instead of uint32_t.

+++ lib/libutee/tee_api_objects.c
@@ -437,7 +437,7 @@ TEE_Result __GP11_TEE_PopulateTransientObject(TEE_ObjectHandle object,
 }
 
 void TEE_InitRefAttribute(TEE_Attribute *attr, uint32_t attributeID,
-			  const void *buffer, uint32_t length)
+			  const void *buffer, size_t length)
 {
 	__utee_check_out_annotation(attr, sizeof(*attr));
